Implement Thick-Skinned Goblin
